pyproject.toml
[project]
name = "phidata"
version = "2.7.8"
description = "Build multi-modal Agents with memory, knowledge and tools."
requires-python = ">=3.7,<4"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  {name = "Ashpreet Bedi", email = "ashpreet@phidata.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "docstring-parser",
  "gitpython",
  "httpx",
  "pydantic",
  "pydantic-settings",
  "python-dotenv",
  "pyyaml",
  "rich",
  "tomli",
  "typer",
  "typing-extensions",
  "openai",
  "duckduckgo-search",
  "phidata",
  "fastapi",
  "uvicorn",
  # Nouvelles dépendances pour le stockage et la vectorisation
  "sqlalchemy",
  "psycopg2-binary",
  "pgvector",
  "pandas",
  "numpy",
  "openpyxl",
  "requests",
  "colorama",  # Ajout de colorama pour le logging coloré
  "python-multipart",  # Ajout pour la gestion des données de formulaire
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "ruff",
    "types-pyyaml",
    "timeout-decorator",
]
docker = [
    "docker"
]
aws = [
    "docker",
    "boto3"
]
k8s = [
    "docker",
    "kubernetes"
]
server = [
    "fastapi",
    "uvicorn",
]
all = [
    "mypy",
    "pytest",
    "ruff",
    "types-pyyaml",
    "docker",
    "boto3",
    "kubernetes",
    "fastapi",
    "uvicorn",
]

[project.scripts]
phi = "phi.cli.entrypoint:phi_cli"

[project.urls]
homepage = "https://phidata.com"
documentation = "https://docs.phidata.com"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["phi*"]

[tool.setuptools.package-data]
phi = ["py.typed"]
include = ["LICENSE"]

[tool.pytest.ini_options]
log_cli = true
testpaths = "tests"

[tool.ruff]
line-length = 120
exclude = ["phienv*", "aienv*"]
# Ignore `F401` (import violations) in all `__init__.py` files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"phi/k8s/app/traefik/crds.py" = ["E501"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["phienv*", "aienv*", "scratch*", "wip*", "tmp*", "cookbook/assistants/examples/*", "phi/assistant/openai/*"]

[[tool.mypy.overrides]]
module = [
  "altair.*",
  "anthropic.*",
  "apify_client.*",
  "arxiv.*",
  "atlassian.*",
  "boto3.*",
  "botocore.*",
  "bs4.*",
  "cassio.*",
  "chonkie.*",
  "chromadb.*",
  "clickhouse_connect.*",
  "clip.*",
  "cohere.*",
  "crawl4ai.*",
  "docker.*",
  "docx.*",
  "duckdb.*",
  "duckduckgo_search.*",
  "email_validator.*",
  "exa_py.*",
  "fastapi.*",
  "firecrawl.*",
  "github.*",
  "google.*",
  "googlesearch.*",
  "groq.*",
  "huggingface_hub.*",
  "jira.*",
  "kubernetes.*",
  "lancedb.*",
  "langchain.*",
  "langchain_core.*",
  "llama_index.*",
  "mem0.*",
  "mistralai.*",
  "mlx_whisper.*",
  "nest_asyncio.*",
  "newspaper.*",
  "numpy.*",
  "ollama.*",
  "openai.*",
  "openbb.*",
  "pandas.*",
  "pgvector.*",
  "PIL.*",
  "pinecone.*",
  "pinecone_text.*",
  "psycopg.*",
  "psycopg2.*",
  "pyarrow.*",
  "pycountry.*",
  "pymongo.*",
  "pypdf.*",
  "pytz.*",
  "qdrant_client.*",
  "rapidocr_onnxruntime.*",
  "replicate.*",
  "requests.*",
  "scrapegraph_py.*",
  "sentence_transformers.*",
  "serpapi.*",
  "setuptools.*",
  "simplejson.*",
  "slack_sdk.*",
  "spider.*",
  "sqlalchemy.*",
  "starlette.*",
  "streamlit.*",
  "tantivy.*",
  "tavily.*",
  "textract.*",
  "timeout_decorator.*",
  "torch.*",
  "tzlocal.*",
  "uvicorn.*",
  "vertexai.*",
  "voyageai.*",
  "wikipedia.*",
  "yfinance.*",
  "youtube_transcript_api.*",
]
ignore_missing_imports = true

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
phidata = "^1.5.0"
openai = "^1.0.0"
duckduckgo-search = "^3.8.5"
pandas = "^2.2.0"
numpy = "^1.26.0"
openpyxl = "^3.1.0"  # Pour le support Excel
googlesearch-python = "^1.3.0"
pycountry = "^24.6.1"
llm-axe = "^0.1.0"  # Ajouter la dépendance llm-axe
